Progress on frdr query integration

+++ database/db_helper.py
@@ -1,5 +1,35 @@
 import django
 import numpy as np
+FIELD_LOOKUPS = ['exact',
+                'iexact',
+                'contains',
+                'icontains',
+                'in',
+                'gt',
+                'gte',
+                'lt',
+                'lte',
+                'startswith',
+                'istartswith',
+                'endswith',
+                'iendswith',
+                'range',
+                'date',
+                'year',
+                'iso_year',
+                'month',
+                'day',
+                'week',
+                'week_day',
+                'iso_week_day',
+                'quarter',
+                'time',
+                'hour',
+                'minute',
+                'second',
+                'isnull',
+                'regex',
+                'iregex']
 
 def get_foreign_keys(model):
     fkeys = []
@@ -8,7 +38,7 @@ def get_foreign_keys(model):
             fkeys.append(col)
     return fkeys
 
-def build_model(model,data):
+def build_model(model,data,replace = False):
 
     data.replace(np.nan,None, inplace = True)
     # Replace foreign key references ids with actual referenced model
@@ -25,6 +55,6 @@ def build_model(model,data):
                 row_dict[col] = val
 
         row_models.append(model(**(row_dict)))
-        
-    model.objects.all().delete()
+    if replace:
+        model.objects.all().delete()
     model.objects.bulk_create(row_models)
